Вынес методы в отдельный модуль, добавил карту методов в methods.__init__.py. Реализовал динамическую настройку окна между посылкой реквостов, получающих информацию о состоянии системы.
