+++ webui/methods/__init__.py
@@ -1,0 +1,6 @@
+from methods.headers import HeadersModel
+
+# add new methods to this map
+methodsMap = {
+	1: HeadersModel()
+}
